chore: rename & move types and utils
