+++ packages/common/src/types/savedCharts.ts
@@ -2,7 +2,7 @@ import assertUnreachable from '../utils/assertUnreachable';
 import { type ViewStatistics } from './analytics';
 import { type ConditionalFormattingConfig } from './conditionalFormatting';
 import { type ChartSourceType } from './content';
-import { type CompactOrAlias } from './field';
+import { type CompactOrAlias, type FieldId } from './field';
 import { type MetricQuery, type MetricQueryRequest } from './metricQuery';
 // eslint-disable-next-line import/no-cycle
 import { type SpaceShare } from './space';
@@ -691,4 +691,37 @@ export type CalculateTotalFromQuery = {
 export type ApiCalculateTotalResponse = {
     status: 'ok';
     results: Record<string, number>;
+};
+
+export type ReplaceableFieldMatchMap = {
+    [fieldId: string]: {
+        fieldId: string;
+        label: string;
+        match: {
+            fieldId: FieldId;
+            fieldLabel: string;
+        } | null;
+        suggestedMatches: Array<{
+            fieldId: FieldId;
+            fieldLabel: string;
+        }>;
+    };
+};
+
+export type ReplaceableCustomFields = {
+    [chartUuid: string]: {
+        uuid: string;
+        label: string;
+        customMetrics: ReplaceableFieldMatchMap;
+    };
+};
+
+export type ReplaceCustomFields = {
+    [chartUuid: string]: {
+        customMetrics: {
+            [customMetricId: string]: {
+                replaceWithFieldId: string;
+            };
+        };
+    };
 };
